// src/app/shared/interfaces/cadastro.interface.ts
  export interface CadastroData {
        CpfCnpj: string;
        telefoneCelular: string;
        cep: string;
        logradouro: string;
        bairro: string;
        cidade: string;
        uf: string;
        numero: string;
        nome: string;
        complemento: string;
        emailInput: string;
  }